Add more Windows detail
Added more details to the Windows section, including nuances with vcpkg and finish_freezing, which are important steps that users may get "lost" on without this additional detail. It is not necessarily important to overload users with %MY_PATH% bits, but to also give them examples from a live working system. In this case my Windows 10 box is very common (I think), so I have used that as the basis for writing the examples. The reader ought to be smart enough to make adaptations for their particular machine and its directory structures.